feat: Add metadata
