+++ app/(application)/application/content/application.tsx
@@ -10,7 +10,6 @@ import {
 import Image from "next/image";
 import notice from "@/images/icon/notice.svg";
 import closed from "@/images/icon/popup/closed.svg";
-
 import ListSection from "./listSection";
 import SubTap from "@/components/tab/subTap";
 import SmallBtn from "@/components/button/basicBtn/smallBtn";
@@ -19,8 +18,31 @@ import Alert from "@/components/alert/alert";
 import ErrorNotice from "@/components/feedback/error";
 
 import { OptionType } from "@/types/components/pulldown";
+import { Metadata } from "next";
 
 const ITEMS_PER_PAGE = 10;
+
+export const metadata: Metadata = {
+  title: "내 지원 내역 | 지원 현황을 확인하세요",
+  description:
+    "지원한 동아리 목록과 합격 여부, 진행 상황을 확인할 수 있는 공간입니다.",
+  openGraph: {
+    title: "내 지원 내역 | 지원 현황을 확인하세요",
+    description:
+      "지원한 동아리 목록과 합격 여부, 진행 상황을 확인할 수 있는 공간입니다.",
+    url: "https://ariari.com/application",
+    siteName: "아리아리",
+    images: [
+      {
+        url: "/logo.svg",
+        width: 1200,
+        height: 630,
+        alt: "아리아리",
+      },
+    ],
+    type: "website",
+  },
+};
 
 const Application = () => {
   const router = useRouter();
