+++ app/(club)/club/management/members/components/invitationFormContent.tsx
@@ -64,7 +64,7 @@ const InvitationFormContent = ({
           </h2>
           <button
             onClick={onClose}
-            className="md:w-5 md:h-5 flex justify-center items-center md:w-7 md:h-7"
+            className="flex justify-center items-center md:w-6 md:h-6"
           >
             <Image
               src={close}
