+++ app/(terms)/terms/club/content/mainSection.tsx
@@ -1,0 +1,49 @@
+"use client";
+
+import HeaderSection from "@/(club)/club/withdrawal/content/headerSection";
+import useResponsive from "@/hooks/useResponsive";
+import { Metadata } from "next";
+import MarkdownSection from '../../content/markdownSection';
+
+export const metadata: Metadata = {
+  title: "운영원칙 | 아리아리",
+  description:
+    "아리아리에서 제공하는 동아리 운영 관련 기본 원칙과 규정을 안내합니다.",
+  openGraph: {
+    title: "운영원칙 | 아리아리",
+    description:
+      "아리아리에서 제공하는 동아리 운영 관련 기본 원칙과 규정을 안내합니다.",
+    url: "https://ariari.com/terms/club",
+    siteName: "아리아리",
+    images: [
+      {
+        url: "/logo.svg",
+        width: 1200,
+        height: 630,
+        alt: "아리아리",
+      },
+    ],
+    type: "website",
+  },
+};
+
+const MainSection = () => {
+  const isMdUp = useResponsive("md");
+
+  return (
+    <div className="mt-[46px] mb-20 md:mb-0">
+      {!isMdUp && <HeaderSection title={"동아리 운영원칙"} />}
+      <div className="hidden gap-[22px] mt-6 md:flex md:flex-col md:mt-8 md:">
+        <h1 className="text-text1 md:text-h1_contents_title">
+          동아리 운영원칙
+        </h1>
+      </div>
+
+      <div className="w-full mt-5 mb-20 rounded-8 bg-background  md:mb-[124px] md:mt-[22px] md:px-6 md:py-[26px] ">
+        <MarkdownSection termType={"CLUB_RULES"} />
+      </div>
+    </div>
+  );
+};
+
+export default MainSection;
